requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
xarray>=0.19.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Geospatial processing
rasterio>=1.2.0
rioxarray>=0.7.0
shapely>=1.8.0
pyproj>=3.2.0

# Earth observation
earthengine-api>=0.1.300
geemap>=0.16.0

# Data handling
netCDF4>=1.5.7
h5netcdf>=0.11.0
dask>=2021.10.0
zarr>=2.10.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
cartopy>=0.20.0
plotly>=5.3.0

# Image processing
Pillow>=8.3.0
imageio>=2.9.0

# Water data
dataretrieval>=0.7  # USGS water data

# Utilities
pyyaml>=5.4.0
joblib>=1.0.0
tqdm>=4.62.0
python-dateutil>=2.8.0

# Development (optional)
pytest>=6.2.0
black>=21.7b0
flake8>=3.9.0
jupyter>=1.0.0
ipykernel>=6.0.0